Turnstile example of Event-B implementation develop by hand but keeping in mind a future automatic code generation that starts from a scxml file. It currently has 2 levels of refinement, and parallel states To do: 1. Add one more level of refinement 2. Add clock 3. Add final state
Note: In the process of making this implemenation I see the need to keep information about the region names that are capture by the yakindu tool.  Currently the scxml code gereration makes use of the regions only for parallel states. It seems that keeping even the main region could be helpfull.